+++ coursemos_spring_boot/src/main/java/com/example/demo/service/CouponService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import com.example.demo.domain.Coupon;
 
 public interface CouponService {
-	List<Coupon> getCouponByCompanyId(int companyId);
+	List<Coupon> getCouponByCompanyId(Long companyId);
 	Coupon getCouponByCouponId(int couponId); 
 	void deleteCoupon(Coupon coupon);
 	void updateCoupon(Coupon coupon);
